added controller classes, service class annotations, and resource not found exception

+++ list-constructor-backend/src/main/java/com/project/listconstructorbackend/service/BaseService.java
@@ -15,13 +15,6 @@ public interface BaseService<T extends BaseEntity> {
         return getRepository().save(t);
     }
 
-    default Optional<T> update(T t) {
-        if (getRepository().existsById(t.getId())) {
-            return Optional.of(getRepository().save(t));
-        }
-        return Optional.empty();
-    }
-
     default Optional<T> get(UUID id) {
         return getRepository().findById(id);
     }
@@ -30,8 +23,19 @@ public interface BaseService<T extends BaseEntity> {
         return getRepository().findAll();
     }
 
-    default void delete(UUID id) {
-        getRepository().deleteById(id);
+    default Optional<T> update(T t) {
+        if (getRepository().existsById(t.getId())) {
+            return Optional.of(getRepository().save(t));
+        }
+        return Optional.empty();
+    }
+
+    default boolean delete(UUID id) {
+        if (getRepository().existsById(id)) {
+            getRepository().deleteById(id);
+            return true;
+        }
+        return false;
     }
 
 }
